transform the signup form into a login form
* delete unused fields in the form
* replace all `signup` ocurrences for `login` in ts and css
* rename all ts and css files that included the word `signup`
